refactor: convert all crawlers into classes with inheritance
