+++ ososedki_dl/crawlers/ososedki/__init__.py
@@ -1,0 +1,19 @@
+from ososedki_dl.crawlers.ososedki.cosplayasian import CosplayAsianCrawler
+from ososedki_dl.crawlers.ososedki.cosplayboobs import CosplayBoobsCrawler
+from ososedki_dl.crawlers.ososedki.cosplayrule34 import CosplayRule34Crawler
+from ososedki_dl.crawlers.ososedki.cosplaythots import CosplayThotsCrawler
+from ososedki_dl.crawlers.ososedki.ocosplay import OCosplayCrawler
+from ososedki_dl.crawlers.ososedki.ososedki import OsosedkiCrawler
+from ososedki_dl.crawlers.ososedki.vipthots import VipThotsCrawler
+from ososedki_dl.crawlers.ososedki.waifubitches import WaifuBitchesCrawler
+
+crawlers = [
+    CosplayAsianCrawler,
+    CosplayBoobsCrawler,
+    CosplayRule34Crawler,
+    CosplayThotsCrawler,
+    OCosplayCrawler,
+    OsosedkiCrawler,
+    VipThotsCrawler,
+    WaifuBitchesCrawler,
+]
