+++ ososedki_dl/crawlers/other/__init__.py
@@ -1,0 +1,18 @@
+from ososedki_dl.crawlers.other.bunkrr_albums import BunkrAlbumsCrawler
+from ososedki_dl.crawlers.other.cosxuxi_club import CosxuxiClubCrawler
+from ososedki_dl.crawlers.other.eromexxx import EromeXXXCrawler
+from ososedki_dl.crawlers.other.fapello_is import FapelloIsCrawler
+from ososedki_dl.crawlers.other.husvjjal_blogspot import \
+    HusvjjalBlogspotCrawler
+from ososedki_dl.crawlers.other.sorrymother import SorryMotherCrawler
+from ososedki_dl.crawlers.other.wildskirts import WildskirtsCrawler
+
+crawlers = [
+    BunkrAlbumsCrawler,
+    CosxuxiClubCrawler,
+    EromeXXXCrawler,
+    FapelloIsCrawler,
+    HusvjjalBlogspotCrawler,
+    SorryMotherCrawler,
+    WildskirtsCrawler
+]
