+++ ososedki_dl/crawlers/other/cosxuxi_club.py
@@ -1,0 +1,83 @@
+"""Downloader for https://cosxuxi.club"""
+
+from pathlib import Path
+from typing import override
+
+from bs4 import BeautifulSoup
+from bs4.element import NavigableString, Tag
+
+from ososedki_dl.crawlers._common import (CrawlerContext, download_media_items,
+                                          fetch_soup)
+from ososedki_dl.crawlers.simple_crawler import SimpleCrawler
+from ososedki_dl.utils import get_final_path
+
+
+class CosxuxiClubCrawler(SimpleCrawler):
+    site_url = "https://cosxuxi.club"
+    base_url: str = ".wp.com/img.nungvl.net/"
+
+    def cosxuxi_club_title_extractor(self, soup: BeautifulSoup) -> str:
+        text_div: Tag | NavigableString | None = soup.find("title")
+        if not text_div:
+            return "Unknown"
+        text: str = text_div.text.strip()
+        title: str = "Unknown"
+        try:
+            if "CosXuxi Club: " in text and " - " in text:
+                title = text.split("CosXuxi Club: ")[1].split(" - ")[0].strip()
+        except IndexError:
+            print(f"ERROR: Could not extract title from '{text}'")
+        return title
+
+    def cosxuxi_club_media_filter(self, soup: BeautifulSoup) -> list[str]:
+        # Find all the images inside the div with the class 'contentme'
+        content_div: Tag | NavigableString | None = soup.find("div", class_="contentme")
+        if not content_div or isinstance(content_div, NavigableString):
+            return []
+
+        return [
+            img.get("src")
+            for img in content_div.find_all("img")
+            if self.base_url in img.get("src")
+        ]
+
+    @override
+    async def download(self, context: CrawlerContext, url: str) -> list[dict[str, str]]:
+        album_url: str = url
+        if album_url.endswith("/"):
+            album_url = album_url[:-1]
+
+        title: str = ""
+        urls: list[str] = []
+
+        while True:
+            soup: BeautifulSoup | None = await fetch_soup(context.session, album_url)
+            if not soup:
+                break
+            page_urls: list[str] = self.cosxuxi_club_media_filter(soup) if soup else []
+            if not page_urls:
+                break
+
+            urls.extend(page_urls)
+
+            if not title:
+                title = self.cosxuxi_club_title_extractor(soup)
+
+            # Check if there is a next page
+            next_page: Tag | NavigableString | None = soup.find(
+                "a", class_="page-numbers", string="Next >"
+            )
+            if isinstance(next_page, NavigableString):
+                next_page = None
+            if not next_page or not next_page.get("href"):
+                break
+            next_page_url: str | list[str] = next_page.get("href", "")
+            if isinstance(next_page_url, list):
+                next_page_url = next_page_url[0]
+            album_url = self.site_url + next_page_url
+
+        album_path: Path = get_final_path(context.download_path, title)
+
+        return await download_media_items(
+            context.session, urls, album_path, context.progress, context.task
+        )
