+++ ososedki_dl/crawlers/simple_crawler.py
@@ -1,0 +1,11 @@
+from abc import ABC, abstractmethod
+
+from ososedki_dl.crawlers._common import CrawlerContext
+
+
+class SimpleCrawler(ABC):
+    site_url: str
+
+    @abstractmethod
+    async def download(self, context: CrawlerContext, url: str) -> list[dict[str, str]]:
+        raise NotImplementedError("Each crawler must implement its own download method")
